requirements.txt
apache-airflow==2.9.3
pandas
numpy
kafka-python
pymysql
scipy